rename module from search to filter
