+++ internal/filter/ast.go
@@ -1,4 +1,4 @@
-package search
+package filter
 
 import (
 	"fmt"
