+++ internal/filter/lexer_test.go
@@ -1,4 +1,4 @@
-package search
+package filter
 
 import (
 	"strings"
