+++ internal/filter/parser_internal_test.go
@@ -1,4 +1,4 @@
-package search
+package filter
 
 import (
 	"fmt"
