+++ internal/filter/token.go
@@ -1,4 +1,4 @@
-package search
+package filter
 
 type Token int
